[hotfix/#195] 파일 구조 변경 및 클릭 이벤트 수정
