+++ app/src/main/java/com/smashing/app/presentation/home/HomeScreen.kt
@@ -1,21 +1,17 @@
 package com.smashing.app.presentation.home
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyListState
@@ -23,7 +19,6 @@ import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Icon
@@ -36,7 +31,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
@@ -44,7 +38,6 @@ import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.layout.positionInWindow
 import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.style.TextAlign
@@ -56,27 +49,16 @@ import androidx.compose.ui.window.PopupProperties
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.smashing.app.R
-import com.smashing.app.R.drawable.ic_bell
-import com.smashing.app.R.drawable.ic_bell_notification
-import com.smashing.app.R.drawable.img_dummy_versus
-import com.smashing.app.core.designsystem.component.button.SmashingBaseButton
 import com.smashing.app.core.designsystem.component.card.MatchingCard
-import com.smashing.app.core.designsystem.component.dropdown.RegionDropdown
-import com.smashing.app.core.designsystem.component.image.UrlImage
 import com.smashing.app.core.designsystem.component.ranking.SmashingRankingItem
 import com.smashing.app.core.designsystem.component.toast.LocalToastTrigger
 import com.smashing.app.core.designsystem.state.MatchingCardState
-import com.smashing.app.core.designsystem.style.SmashingBtnColor
 import com.smashing.app.core.designsystem.style.TierInfoStyle
-import com.smashing.app.core.designsystem.style.getMatchButtonColor
-import com.smashing.app.core.designsystem.style.getMatchButtonTitle
 import com.smashing.app.core.designsystem.style.toTierInfoStyle
 import com.smashing.app.core.designsystem.theme.SmashingTheme
 import com.smashing.app.core.extension.noRippleClickable
-import com.smashing.app.core.util.ProfileImageProvider
 import com.smashing.app.core.util.ScrollStateHolder
 import com.smashing.app.core.util.bottomBarNestedScrollConnection
-import com.smashing.app.data.model.matching.AcceptedMatching
 import com.smashing.app.data.model.my.ActiveUserProfile
 import com.smashing.app.data.model.rank.UserRank
 import com.smashing.app.data.model.search.SearchMainItemModel
@@ -84,9 +66,10 @@ import com.smashing.app.data.type.GameResultStatusType
 import com.smashing.app.data.type.GenderType
 import com.smashing.app.data.type.SportType
 import com.smashing.app.data.type.TierType
+import com.smashing.app.presentation.home.component.CloseMatching
 import com.smashing.app.presentation.home.component.HomeDropdown
+import com.smashing.app.presentation.home.component.HomeTopBar
 import com.smashing.app.presentation.home.component.RecommendedInfoPopup
-import com.smashing.app.presentation.home.component.SportsTierChip
 import kotlinx.collections.immutable.toImmutableList
 
 
@@ -111,6 +94,7 @@ fun HomeRoute(
         gameId: String,
         isFirstAttempt: Boolean,
     ) -> Unit,
+    navigateToMyProfile: () -> Unit,
     updateBottomBar: (Boolean) -> Unit,
     modifier: Modifier = Modifier,
     viewModel: HomeViewModel = hiltViewModel(),
@@ -154,6 +138,7 @@ fun HomeRoute(
         navigateToSearch = navigateToSearch,
         navigateToSubmit = navigateToSubmit,
         navigateToConfirm = navigateToConfirm,
+        navigateToMyProfile = navigateToMyProfile,
         onSportsChipClick = viewModel::fetchSelectSportProfile,
         updateBottomBar = updateBottomBar,
         recommendedUserListState = recommendedUserListState,
@@ -183,6 +168,7 @@ private fun HomeScreen(
         gameId: String,
         isFirstAttempt: Boolean,
     ) -> Unit,
+    navigateToMyProfile: () -> Unit,
     onSportsChipClick: (String) -> Unit,
     updateBottomBar: (Boolean) -> Unit,
     modifier: Modifier = Modifier,
@@ -505,7 +491,13 @@ private fun HomeScreen(
                             tier = ranker.tier,
                             lp = ranker.lp,
                             userId = ranker.userId,
-                            onClick = { navigateToUserProfile(ranker.userId) },
+                            onClick = {
+                                if (ranker.userId != uiState.activeUserProfile.profileId) {
+                                    navigateToUserProfile(ranker.userId)
+                                } else {
+                                    navigateToMyProfile()
+                                }
+                            },
                         )
                     }
                 }
@@ -547,214 +539,6 @@ private fun HomeScreen(
                 )
             }
         }
-    }
-}
-
-@Composable
-private fun HomeTopBar(
-    userRegion: String,
-    userSport: SportType,
-    userTier: TierType,
-    onClickRegion: (String) -> Unit,
-    onChangeRegion: () -> Unit,
-    onClickSportChip: () -> Unit,
-    onClickNotice: () -> Unit,
-    modifier: Modifier = Modifier,
-    isNotice: Boolean = false,
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(
-                horizontal = 16.dp,
-            ),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        RegionDropdown(
-            selectedItem = userRegion,
-            items = listOf(
-                userRegion
-            ).toImmutableList(),
-            onClick = onClickRegion,
-            onRegionChange = onChangeRegion,
-            isDivide = true,
-            modifier = Modifier
-                .padding(vertical = 12.dp)
-        )
-
-        Spacer(modifier = Modifier.weight(1f))
-
-        SportsTierChip(
-            sportType = userSport,
-            tierType = userTier,
-            onClick = onClickSportChip,
-        )
-
-        Spacer(modifier = Modifier.width(12.dp))
-
-        if (!isNotice) {
-            Icon(
-                imageVector = ImageVector.vectorResource(ic_bell),
-                contentDescription = null,
-                tint = SmashingTheme.colors.iconPrimary,
-                modifier = Modifier
-                    .noRippleClickable(
-                        onClick = onClickNotice,
-                    )
-            )
-        } else {
-            Icon(
-                imageVector = ImageVector.vectorResource(ic_bell_notification),
-                contentDescription = null,
-                tint = SmashingTheme.colors.iconPrimary,
-                modifier = Modifier
-                    .noRippleClickable(
-                        onClick = onClickNotice,
-                    )
-            )
-        }
-    }
-}
-
-@Composable
-private fun CloseMatching(
-    myNickname: String,
-    myProfileId: String,
-    onClick: (AcceptedMatching) -> Unit,
-    modifier: Modifier = Modifier,
-    matchedUser: AcceptedMatching? = null,
-    navigateToSearch: () -> Unit,
-) {
-    Column(
-        modifier = modifier
-            .fillMaxWidth()
-            .background(
-                color = SmashingTheme.colors.bgSurface,
-                shape = RoundedCornerShape(8.dp),
-            )
-            .padding(
-                horizontal = 16.dp,
-            )
-            .padding(
-                top = 22.dp,
-                bottom = 24.dp,
-            ),
-    ) {
-        if (matchedUser != null) {
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-            ) {
-                Image(
-                    painter = painterResource(id = img_dummy_versus),
-                    contentDescription = null,
-                    modifier = Modifier
-                        .size(
-                            width = 100.dp,
-                            height = 108.dp,
-                        )
-                        .align(Alignment.Center),
-                )
-
-                MatchedUserItem(
-                    userId = myProfileId,
-                    nickname = myNickname,
-                    modifier = Modifier.align(Alignment.CenterStart)
-                )
-
-                MatchedUserItem(
-                    userId = matchedUser.userId,
-                    nickname = matchedUser.nickname,
-                    modifier = Modifier.align(Alignment.CenterEnd)
-                )
-            }
-
-        } else {
-            Text(
-                text = stringResource(R.string.home_no_matching),
-                style = SmashingTheme.typography.md.medium16,
-                color = SmashingTheme.colors.txtTertiary,
-                textAlign = TextAlign.Center,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(
-                        top = 30.dp,
-                        bottom = 2.dp,
-                    )
-            )
-        }
-
-        Spacer(modifier = Modifier.height(22.dp))
-
-        if (matchedUser != null) {
-            SmashingBaseButton(
-                text = matchedUser.resultStatus.getMatchButtonTitle(),
-                textStyle = SmashingTheme.typography.md.medium16,
-                onClick = { onClick(matchedUser) },
-                buttonColor = matchedUser.resultStatus.getMatchButtonColor(),
-                contentPadding = PaddingValues(
-                    vertical = 9.dp,
-                ),
-                modifier = Modifier
-                    .fillMaxWidth(),
-                shape = RoundedCornerShape(8.dp),
-                isRippleEnabled = false,
-            )
-        } else {
-            SmashingBaseButton(
-                modifier = Modifier.fillMaxWidth()
-                    .padding(bottom = 28.dp),
-                text = "매칭 탐색하러 가기",
-                textStyle = SmashingTheme.typography.md.medium16,
-                onClick = navigateToSearch,
-                buttonColor = SmashingBtnColor(
-                    backgroundColor = SmashingTheme.colors.btnBgPrimary300,
-                    textColor = SmashingTheme.colors.txtEmphasis,
-                    disabledBackgroundColor = SmashingTheme.colors.btnBgPrimary300,
-                    disabledTextColor = SmashingTheme.colors.txtEmphasis,
-                ),
-                contentPadding = PaddingValues(vertical = 9.dp),
-                shape = RoundedCornerShape(8.dp),
-                isRippleEnabled = false,
-            )
-        }
-    }
-}
-
-@Composable
-private fun MatchedUserItem(
-    userId: String,
-    nickname: String,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier
-            .width(IntrinsicSize.Max),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Box(
-            modifier = Modifier
-                .padding(
-                    horizontal = 28.dp,
-                )
-        ) {
-            UrlImage(
-                placeholderDrawable = ProfileImageProvider.getTempImg(nickname),
-                modifier = Modifier
-                    .height(64.dp)
-                    .aspectRatio(1f)
-                    .clip(CircleShape),
-            )
-        }
-
-        Spacer(modifier = Modifier.height(4.dp))
-
-        Text(
-            text = nickname,
-            style = SmashingTheme.typography.sm.medium14,
-            color = SmashingTheme.colors.txtMuted,
-        )
     }
 }
 
@@ -886,12 +670,13 @@ private fun HomeScreenPreview() {
         navigateToRanking = {},
         navigateToMatchingAccepted = {},
         navigateToUserProfile = {},
-        onSportsChipClick = {},
         navigateToSportAdd = {},
         navigateToSearch = {},
-        updateBottomBar = {},
         navigateToSubmit = { _, _, _, _ -> },
         navigateToConfirm = { _, _, _ -> },
+        navigateToMyProfile = {},
+        onSportsChipClick = {},
+        updateBottomBar = {},
     )
 }
 
@@ -933,9 +718,10 @@ private fun HomeScreenEmptyValuePreview() {
         navigateToUserProfile = {},
         navigateToSportAdd = {},
         navigateToSearch = {},
-        onSportsChipClick = {},
-        updateBottomBar = {},
+        navigateToMyProfile = {},
         navigateToSubmit = { _, _, _, _ -> },
         navigateToConfirm = { _, _, _ -> },
+        onSportsChipClick = {},
+        updateBottomBar = {},
     )
 }
