+++ app/src/main/java/com/smashing/app/presentation/home/navigation/HomeNavigation.kt
@@ -14,6 +14,7 @@ import com.smashing.app.presentation.home.regionchange.RegionChangeRoute
 import com.smashing.app.presentation.matching.navigation.navigateToMatching
 import com.smashing.app.presentation.matching.type.MatchingType
 import com.smashing.app.presentation.notice.navigation.navigateToNotice
+import com.smashing.app.presentation.profile.navigation.navigateToMyProfile
 import com.smashing.app.presentation.profile.navigation.navigateToUserProfile
 import com.smashing.app.presentation.ranking.navigation.navigateToRanking
 import com.smashing.app.presentation.region.navigation.getRegionResult
@@ -77,6 +78,7 @@ fun NavGraphBuilder.homeGraph(
                         isFirstAttempt = isFirstAttempt,
                     )
                 },
+                navigateToMyProfile = navController::navigateToMyProfile,
                 updateBottomBar = updateBottomBar,
             )
         }
